semantic-search-api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
numpy
pydantic
